requirements.txt
Flask==2.3.3
deep-translator==1.11.4
requests==2.31.0
python-docx==1.2.0
pydub==0.25.1
